backend_python/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.4.1
openai==1.13.3
firebase-admin==6.4.0
python-pptx==0.6.21
fpdf2==2.7.8
reportlab==4.1.0
python-multipart==0.0.9
pydantic==2.6.3
aiofiles==23.2.1
requests==2.31.0
